Dont restart checkout_flow for completed orders
No need to destroy all shipments or set order back to first checkout
state. Currently shipments for completed orders are managed by
OrderInventory#veriry

Conflicts:
	core/spec/models/spree/order_spec.rb

+++ core/app/models/spree/order.rb
@@ -508,7 +508,7 @@ module Spree
     # to delivery again so that proper updated shipments are created.
     # e.g. customer goes back from payment step and changes order items
     def ensure_updated_shipments
-      if shipments.any?
+      if shipments.any? && !self.completed?
         self.shipments.destroy_all
         restart_checkout_flow
       end
